+++ src/@batch-flask/ui/file/details/code-file-viewer/code-file-viewer.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnChanges } from "@angular/core";
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input, OnChanges } from "@angular/core";
 import { EditorConfig } from "@batch-flask/ui/editor";
 import { FileLoader } from "@batch-flask/ui/file/file-loader";
 import { File } from "@batch-flask/ui/file/file.model";
@@ -11,6 +11,7 @@ const maxSize = 100000; // 100KB
 @Component({
     selector: "bl-code-file-viewer",
     templateUrl: "code-file-viewer.html",
+    changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CodeFileViewerComponent implements OnChanges {
     @Input() public fileLoader: FileLoader;
@@ -28,6 +29,8 @@ export class CodeFileViewerComponent implements OnChanges {
         },
     };
 
+    constructor(private changeDetector: ChangeDetectorRef) {}
+
     public ngOnChanges(changes) {
         this._loadImage();
     }
@@ -36,6 +39,8 @@ export class CodeFileViewerComponent implements OnChanges {
         this.fileTooLarge = false;
         this.loadingStatus = LoadingStatus.Loading;
         this.fileNotFound = false;
+        this.changeDetector.markForCheck();
+
         this.fileLoader.getProperties().subscribe({
             next: (file: File) => {
                 this.file = file;
@@ -49,7 +54,9 @@ export class CodeFileViewerComponent implements OnChanges {
                 this.fileLoader.content().subscribe((result) => {
                     this.value = result.content.toString();
                     this.loadingStatus = LoadingStatus.Ready;
+                    this.changeDetector.markForCheck();
                 });
+                this.changeDetector.markForCheck();
             },
             error: (error) => null,
         });
