+++ src/@batch-flask/ui/file/details/file-content/file-content.component.spec.ts
