+++ src/@batch-flask/ui/file/details/file-content/file-content.html
