+++ src/@batch-flask/ui/file/details/file-content/index.ts
@@ -1,0 +1,1 @@
+export * from "./file-content.component";
