+++ src/@batch-flask/ui/file/details/index.ts
@@ -1,3 +1,3 @@
 export * from "./file-details.module";
 export * from "./file-details-view";
-export * from "./file-content.component";
+export * from "./file-content";
